changes: delete redux theme endpoints, rename api.ts --> theme.api.ts
